pyproject.toml
[build-system]
requires = ["setuptools>=64"]
build-backend = "setuptools.build_meta"

[project]
name = "lume-visualizations"
version = "0.1.0"
description = "Marimo beam-monitor dashboards for the LCLS CU injector virtual accelerator"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "caproto>=1.3",
    "marimo>=0.22,<1",
    "matplotlib>=3.8",
    "numpy>=1.26",
    "pyepics>=3.5",
    "virtual-accelerator @ git+https://github.com/pluflou/virtual-accelerator.git@injector_pv_transform",
]
# "virtual-accelerator @ git+https://github.com/slaclab/virtual-accelerator.git@09d1be14d44e682d5b7a08d2ae46232168828091",


[project.scripts]
lume-fake-epics-ioc = "lume_visualizations.fake_epics_ioc:main"
lume-quad-scan = "lume_visualizations.cli:quad_scan_main"
lume-live-monitor = "lume_visualizations.cli:live_stream_main"

[tool.setuptools.packages.find]
include = ["lume_visualizations"]
namespaces = false
